기믹 컴포넌트 설계
1. 몇 턴안에 얼만큼의 데미지를 입히지 못하면 큰 피해를 입히기

2. 여러스택의 버프를 가지고 있으며, 이 스택이 있을때 모든 피해를 1씩만 입고 스택을 하나씩 까서 연타공격이 있는 카드를 사용하게 하는 기믹

3. 플레이어에게 드로우를 방해하는 카드를 삽입하고, 주기적으로 데미지를 1받는 기믹

4. 전투중에 스택을 쌓으며, 체력이 0이되어 사망할 때, 스택만큼 엄청난 데미지를 플레이어에게 입히는 기믹

5. 몇턴간 잠들어 있고, 시간이 지나면 깨어나서 엄청나게 강해지는 기믹

6. 하수인을 소환하여 하수인이 있을때, 버프를 얻고 힐을 하는 기믹

+++ SlayTheChampions/Source/SlayTheChampions/Public/Unit/CombatTypes.h
@@ -36,7 +36,7 @@ struct SLAYTHECHAMPIONS_API FIntent
     GENERATED_BODY()
 
     UPROPERTY(BlueprintReadOnly) EIntentKind Kind = EIntentKind::Unknown;
-    UPROPERTY(BlueprintReadOnly) float       Value = 0.f;  // 예상 데미지 or 블록량
+    UPROPERTY(BlueprintReadOnly) int32       Value = 0;  // 예상 데미지 or 블록량
     UPROPERTY(BlueprintReadOnly) int32       Hits = 1;
     UPROPERTY(BlueprintReadOnly) TWeakObjectPtr<AUnit> Target;
     UPROPERTY(BlueprintReadOnly) FText       DisplayText;
@@ -67,7 +67,7 @@ struct SLAYTHECHAMPIONS_API FEnemyAction
     GENERATED_BODY()
 
     UPROPERTY(EditAnywhere, BlueprintReadOnly) EIntentKind  IntentKind = EIntentKind::Attack;
-    UPROPERTY(EditAnywhere, BlueprintReadOnly) float        Value = 0.f;   // 데미지 or 블록량
+    UPROPERTY(EditAnywhere, BlueprintReadOnly) int32        Value = 0;   // 데미지 or 블록량
     UPROPERTY(EditAnywhere, BlueprintReadOnly) int32        Hits = 1;
     UPROPERTY(EditAnywhere, BlueprintReadOnly) ETargetType  TargetType = ETargetType::SingleEnemy;
     UPROPERTY(EditAnywhere, BlueprintReadOnly) float        Weight = 1.f;   // Weighted 모드에서만 사용
